Version control support for Lambda resources
If a Lambda function code is under version control the commit hash will be added
to the name of the deployment package that is uploaded to S3. This allows you
to run a successful update operation when the version of your lambda changes,
but nothing else has changed in the associated layer. Otherwise CF does not
detect a change in the template and produces an error.